deploy/requirements.txt
fastapi
uvicorn
pydantic
torch
torchvision
pyyaml
python-multipart
Pillow
